fix: index exports
