--- packages/frames.js/src/middleware/images-worker/next/index.ts
@@ -1,1 +1,0 @@
-export * from "./createImagesWorker";
